Dynamically compile spec'd __init__ functions
init_func_generator() previously used a single argument-less `init` base
function coupled with a locals() hack to build spec-aware __init__()
functions. This worked well, but profiling indicated that we were spending a
good amount of time calling locals() and dict.copy()'ing its result.

We now dynamically generate unique, per-spec __init__() functions using
compile(..., 'exec'). This results in much faster runtime performance at the
expense of a little compile-time complexity.

The benchmark indicates this change results in ~10% faster decoding (for both
the native and cybin implementations).

Also, because this code no longer requires version-specific code paths, it can
be moved out of _compat.py and into thrift.py.
